Database: suppress warning caused by migration 004
```
Unhandled exception. System.InvalidOperationException: An error was generated for warning 'Microsoft.EntityFrameworkCore.Migrations.NonTransactionalMigrationOperationWarning': The migration operation 'PRAGMA foreign_keys = 0;
' from migration '_004' cannot be executed in a transaction. If the app is terminated or an unrecoverable error occurs while this operation is being executed then the migration will be left in a partially applied state and would need to be reverted manually before it can be applied again. Create a separate migration that contains just this operation. This exception can be suppressed or logged by passing event ID 'RelationalEventId.NonTransactionalMigrationOperationWarning' to the 'ConfigureWarnings' method in 'DbContext.OnConfiguring' or 'AddDbContext'.
   at Microsoft.EntityFrameworkCore.Diagnostics.EventDefinition`2.Log[TLoggerCategory](IDiagnosticsLogger`1 logger, TParam1 arg1, TParam2 arg2)
   at Microsoft.EntityFrameworkCore.Diagnostics.RelationalLoggerExtensions.NonTransactionalMigrationOperationWarning(IDiagnosticsLogger`1 diagnostics, IMigrator migrator, Migration migration, MigrationCommand command)
   at Microsoft.EntityFrameworkCore.Migrations.Internal.Migrator.<>c__DisplayClass19_2.<GetMigrationCommandLists>b__2()
   at Microsoft.EntityFrameworkCore.Migrations.Internal.Migrator.Migrate(String targetMigration)
   at Microsoft.EntityFrameworkCore.RelationalDatabaseFacadeExtensions.Migrate(DatabaseFacade databaseFacade)
   at Booski.Database.Migrate() in /var/home/ducky/.local/src/booski/src/Booski/Database.cs:line 17
   at Booski.Program.Main(String[] args) in /var/home/ducky/.local/src/booski/src/Booski/Program.cs:line 111
   at Booski.Program.<Main>(String[] args)
```